Updated WP_Codebird::_callApi() to support _buildMultipart()
This fixes issue #6. Added support for _buildMultipart() which posts images to twitter correctly. Until we update Codebird on wpcom, this will allow posting of images. Essentially, it just looks to see if it's a multipart posting and, if so, builds the multipart post body (with the content-disposition tags and boundaries). It also set's the headers correctly.